requirements.txt
pyTelegramBotAPI==4.6.0
python-dotenv==0.20.0
